browser_automation/requirements.txt
# Core dependencies
playwright==1.41.2
langchain==0.1.12
langchain-community==0.0.29
langchain-openai==0.0.8
tiktoken==0.6.0
streamlit==1.32.2

# Additional dependencies
python-dotenv==1.0.1
lxml==5.4.0
beautifulsoup4==4.13.4
pydantic==2.6.4
